Some more refactoring on types + added some comments + remove unnecessary headers in XBDM's pch

+++ OpenNeighborhood/src/Elements/Element.cpp
@@ -24,7 +24,7 @@ void Element::OnRender()
 {
 	auto texture = TextureManager::GetTexture(m_TextureName);
 
-	if (ImGui::ImageButtonWithText((void*)(intptr_t)texture->GetTextureID(), ImVec2((float)texture->GetWidth(), (float)texture->GetHeight()), ImVec2((float)m_Width, (float)m_Height), m_Label.c_str(), ImVec2(m_Padding, m_Padding)))
+	if (ImGui::ImageButtonWithText((void*)(intptr_t)texture->GetTextureID(), ImVec2(texture->GetWidth(), texture->GetHeight()), ImVec2(m_Width, m_Height), m_Label.c_str(), ImVec2(m_Padding, m_Padding)))
 	{
 		if (ImGui::IsMouseDoubleClicked(ImGuiMouseButton_Left))
 			OnClick();
